Fix Windows debug build caused by spdlog (#251)
### Fix Windows debug build caused by spdlog

### Linked issues

### Describe the reason for the change.

This commit fixes the windows debug build

### Summarize your change.

When spdlog is built in debug, the 'd' suffix is added to the resulting
static lib by the spdlog build system.
This 'd' suffix was added to the spdlog.cmake and I took the opportunity
to do some cleanup.
Also, the CMAKE_BUILD_TYPE was NOT passed to the spdlog build on
Mac+Linux. I fixed that too.

### Describe what you have tested and on which operating system.

Tested on all platforms in both debug and release

Signed-off-by: Bernard Laberge <bernard.laberge@autodesk.com>
Signed-off-by: Inouk Babin-Ruel <inouk.babin.ruel@autodesk.com>