feature/vfx2023 update to vs2022 and boost to 1 80 (#250)
<!--
Thanks for your contribution! Please read this comment in its entirety.
It's quite important.
When a contributor merges the pull request, the title and the
description will be used to build the merge commit!

It should be in the following format:

[ 12345: Summary of the changes made ] Where 12345 is the corresponding
Github Issue

OR

[ Summary of the changes made ] If it's solving something trivial, like
fixing a typo.
-->

<!--
Link the Issue(s) this Pull Request is related to.

Each PR should link to at least one issue, in the form:

Use one line for each Issue. This allows auto-closing the related issue
when the fix is merged.

Fixes #12345
Fixes #54345
-->

Fixes #SG-31056

VS2022

Updating to VS2022 and the compiler toolchain V143

VFX 2023 platform update

Windows 11: RV boot start (on Debug)
Both Windows Build Configs built correctly

attention during the review.

- PySide2 in a new Win11 machine with only VS2022 was able to build. RV
also built

---------

Signed-off-by: Inouk Babin-Ruel <inouk.babin.ruel@autodesk.com>