feature/vfx2023 update to vs2022 and boost to 1 80 (#250)
<!--
Thanks for your contribution! Please read this comment in its entirety.
It's quite important.
When a contributor merges the pull request, the title and the
description will be used to build the merge commit!

### Pull Request TITLE

It should be in the following format:

[ 12345: Summary of the changes made ] Where 12345 is the corresponding
Github Issue

OR

[ Summary of the changes made ] If it's solving something trivial, like
fixing a typo.
-->

### Linked issues
<!--
Link the Issue(s) this Pull Request is related to.

Each PR should link to at least one issue, in the form:

Use one line for each Issue. This allows auto-closing the related issue
when the fix is merged.

Fixes #12345
Fixes #54345
-->

Fixes #SG-31056

### VFX 2023 Update: VS2022 with V143, Boost 1.80 and PySide2 with
VS2022

### Summarize your change.

Updating to VS2022 and the compiler toolchain V143

### Describe the reason for the change.

VFX 2023 platform update

### Describe what you have tested and on which operating system.

Windows 11: RV boot start (on Debug)
Both Windows Build Configs built correctly

### Add a list of changes, and note any that might need special
attention during the review.

- PySide2 in a new Win11 machine with only VS2022 was able to build. RV
also built

### If possible, provide screenshots.

---------

Signed-off-by: Inouk Babin-Ruel <inouk.babin.ruel@autodesk.com>