Fix segfault on installed RHEL9 release build (#262)
### Fix segfault when running an installed RHEL9 build

### Summarize your change.

This branch will cause RV to use the system OpenSSL version when it
matches the Python OpenSSL version and to use an installed OpenSSL when
it does not.

I also noticed the python `CMakeFiles` was using the `${target}`
variable to reference the OpenSSL paths. This happened to work because
the python `CMakeFiles` did not set `${target}` and the OpenSSL does
which was the `CMakeFiles` processed just before python. But this was
very lucky -- if `${target}` had been set to anything else before
reaching those lines it would break the build. So I exported the correct
values from the OpenSSL files and used those within Python.

### Describe the reason for the change.

When running an installed build on Rocky9 -- or any Linux distro where
the system OpenSSL version does not match the one used by the Python
build -- OpenRV might segfault on startup.

The reason for this is that Python is complied with a specific OpenSSL
version. We do not install OpenSSL into the libs directory like we do
for other libs because RHEL7 and RHEL8 both have the same version of
OpenSSL as Python (1.1.1) but on RHEL8 it has slightly different symbols
in (which is very annoying). So the workaround here is to just let them
use the system OpenSSL library.

This workaround however breaks RHEL9 since it has completely different
version of OpenSSL than Python. So in this case we need to use Python's
OpenSSL version, but we can't install it into libs without breaking
RHEL8.../

So we need a workaround so RHEL7, 8 and 9 can all work. The only way I
could find of doing this is to install OpenSSL into its own subdirectory
under libs. When loading RV (or any of the other of its binaries) we
will conditionally set the LD_LIBRARY_PATH to point to this library when
the system OpenSSL version does not match Python's.

### Describe what you have tested and on which operating system.

Built successfully on RHEL 7, 8 and 9.

Signed-off-by: Roger Nelson <roger.nelson@autodesk.com>