VFX 2023: Updating Python short version number (#274)
<!--
Thanks for your contribution! Please read this comment in its entirety.
It's quite important.
When a contributor merges the pull request, the title and the
description will be used to build the merge commit!

### Pull Request TITLE

It should be in the following format:

[ 12345: Summary of the changes made ] Where 12345 is the corresponding
Github Issue

OR

[ Summary of the changes made ] If it's solving something trivial, like
fixing a typo.
-->
### Fixing build due to Python update and OCIO merge

### Linked issues
N/A
<!--
Link the Issue(s) this Pull Request is related to.

Each PR should link to at least one issue, in the form:

Use one line for each Issue. This allows auto-closing the related issue
when the fix is merged.

Fixes #12345
Fixes #54345
-->

### Summarize your change.
Fixing the build since OCIO looks for python3.9

### Describe the reason for the change.
The build is broken

### Describe what you have tested and on which operating system.
Windows, Mac & Linux

### Add a list of changes, and note any that might need special
attention during the review.
- Refactoring the code to avoid code duplication and sync'ing the short
number with all other version numbers in this file

### If possible, provide screenshots.

Signed-off-by: Inouk Babin-Ruel <inouk.babin.ruel@autodesk.com>