Fixes OpenRV outputs (#283)
<!--
Thanks for your contribution! Please read this comment in its entirety.
It's quite important.
When a contributor merges the pull request, the title and the
description will be used to build the merge commit!

### Pull Request TITLE

It should be in the following format:

[ 12345: Summary of the changes made ] Where 12345 is the corresponding
Github Issue

OR

[ Summary of the changes made ] If it's solving something trivial, like
fixing a typo.
-->

### Linked issues
<!--
Link the Issue(s) this Pull Request is related to.

Each PR should link to at least one issue, in the form:

Use one line for each Issue. This allows auto-closing the related issue
when the fix is merged.

Fixes #12345
Fixes #54345
-->

N/A

### Summarize your change.

Reverts a change introduced when we introduced the logger where lines
were printed without the newline. Now, we tell SPDLOG to not add an
extra EOL character and keep the one that comes from OpenRV's internal.

### Describe the reason for the change.

Callstacks, and other similar wall of texts, were unreadable.

### Describe what you have tested and on which operating system.

Tested on MacOS, but should not be platform-specific

### Add a list of changes, and note any that might need special
attention during the review.

Reviewing commits per commit is better since the first one is the
clang-format one.

### If possible, provide screenshots.

N/A

---------

Signed-off-by: Kerby Geffrard <kerby.geffrard@autodesk.com>