Another approach to SPDLOG's newline issues (#286)
<!--
Thanks for your contribution! Please read this comment in its entirety.
It's quite important.
When a contributor merges the pull request, the title and the
description will be used to build the merge commit!

### Pull Request TITLE

It should be in the following format:

[ 12345: Summary of the changes made ] Where 12345 is the corresponding
Github Issue

OR

[ Summary of the changes made ] If it's solving something trivial, like
fixing a typo.
-->

### Linked issues
<!--
Link the Issue(s) this Pull Request is related to.

Each PR should link to at least one issue, in the form:

Use one line for each Issue. This allows auto-closing the related issue
when the fix is merged.

Fixes #12345
Fixes #54345
-->

N/A

### Summarize your change.

Reverts the change in #283 as it wasn´t quite right, instead, uses the
mechanism provided by SPDLOG to override the SPDLOG_EOL define at build
time. Note: There are other configs in that file that we might want to
tweak!

Also, I'm trimming the extra white space in the log file at the
beginning of the line if the line got sent with a space between the
"INFO:/ERROR:/..." and the output line.

### Describe the reason for the change.

@savoiei reported that my solution wasn´t optimal on the logger file.
There were extra new lines. It seems like setting SPDLOG_EOL before
importing <spdlog/...> wasn´t enough.

### Describe what you have tested and on which operating system.

Tested on Rocky 9

### Add a list of changes, and note any that might need special
attention during the review.

### If possible, provide screenshots.

Signed-off-by: Kerby Geffrard <kerby.geffrard@autodesk.com>