Create OpenEXR patch to fill invalid lines and tiles with black (#291)
### Patch OpenEXR to fill invalid scanlines and tiles with black

### Summarize your change.

This branch add a patch to OpenEXR after we download it. The patch will
fill any scanlines or tiles that it fails to read with black.

### Describe the reason for the change.

This was a feature that was lost when migrating from RV to OpenRV.
During the migration dependencies were removed directly from the repo
and instead fetched via CMake's `ExternalProject_Add` functionality.
However, when replacing the OpenEXR code we were not aware it had been
modified.

The correct fix for this is to do it in the API directly (or use OIIO to
read OpenEXR which already does exactly that). But until that can be
done, we'll just take the changes from the old system and create a patch
file from them which can be integrated into CMake's `PATCH_COMMAND` step
with `ExternalProject_Add.

### Describe what you have tested and on which operating system.

Mac M1 Ventura 13.1

Signed-off-by: Roger Nelson <roger.nelson@autodesk.com>