Keep RvConsoleWindow output lines intact (#293)
<!--
Thanks for your contribution! Please read this comment in its entirety.
It's quite important.
When a contributor merges the pull request, the title and the
description will be used to build the merge commit!

### Pull Request TITLE

It should be in the following format:

[ 12345: Summary of the changes made ] Where 12345 is the corresponding
Github Issue

OR

[ Summary of the changes made ] If it's solving something trivial, like
fixing a typo.
-->

### Linked issues
<!--
Link the Issue(s) this Pull Request is related to.

Each PR should link to at least one issue, in the form:

Use one line for each Issue. This allows auto-closing the related issue
when the fix is merged.

Fixes #12345
Fixes #54345
-->

N/A

### Summarize your change.

Misc tweaks on the RvConsoleWindow so we don't lose the first characters
of a line without type (ex: INFO: )

### Describe the reason for the change.

While debugging an issue I had working on a Mu package, I was having the
first letters of my backtrace lines stripped. It's because they don't
have the logger type prefix in their lines.

### Describe what you have tested and on which operating system.

Tested on MacOS. 

### Add a list of changes, and note any that might need special
attention during the review.

N/A

### If possible, provide screenshots.

Before:


![image](https://github.com/AcademySoftwareFoundation/OpenRV/assets/6785406/ee707073-aa46-4214-a9e5-7b26e0c336cb)

After: 


![image](https://github.com/AcademySoftwareFoundation/OpenRV/assets/6785406/d0b7205a-a232-48de-9248-8311d3fc9a91)

Signed-off-by: Kerby Geffrard <kerby.geffrard@autodesk.com>