Avoid sending empty headers and cookie to FFMPEG (#296)
<!--
Thanks for your contribution! Please read this comment in its entirety.
It's quite important.
When a contributor merges the pull request, the title and the
description will be used to build the merge commit!

### Pull Request TITLE

It should be in the following format:

[ 12345: Summary of the changes made ] Where 12345 is the corresponding
Github Issue

OR

[ Summary of the changes made ] If it's solving something trivial, like
fixing a typo.
-->

### Linked issues
<!--
Link the Issue(s) this Pull Request is related to.

Each PR should link to at least one issue, in the form:

Use one line for each Issue. This allows auto-closing the related issue
when the fix is merged.

Fixes #12345
Fixes #54345
-->

N/A

### Summarize your change.

1) Avoid submitting cookies and headers parameters to FFMPEG if they are
empty strings... it doesn't like it.
`No trailing CRLF found in HTTP header` -> empty header. 

2) Use the proper delimiter of `\r\n` for headers because that's the
spec.

### Describe the reason for the change.

There was one warning line per web media loaded. 

### Describe what you have tested and on which operating system.

Tested on MacOS, but not platform-specific. 

### Add a list of changes and note any needing special attention during
the review.

The change is in 44c9bdc3ea77ff1a3dba83b7beeff130a8db624b, the other
commit is just automatic formatting from my editor.

### If possible, provide screenshots.

N/A

---------

Signed-off-by: Kerby Geffrard <kerby.geffrard@autodesk.com>