Add script variant of windows reg files for rvlink and rvpush (#302)
<!--
Thanks for your contribution! Please read this comment in its entirety.
It's quite important.
When a contributor merges the pull request, the title and the
description will be used to build the merge commit!

### Pull Request TITLE

It should be in the following format:

[ 12345: Summary of the changes made ] Where 12345 is the corresponding
Github Issue

OR

[ Summary of the changes made ] If it's solving something trivial, like
fixing a typo.
-->

### Linked issues
<!--
Link the Issue(s) this Pull Request is related to.

Each PR should link to at least one issue, in the form:

Use one line for each Issue. This allows auto-closing the related issue
when the fix is merged.

Fixes #12345
Fixes #54345
-->

N/A

### Summarize your change.

Provide a script that sets the required registry entries for rvpush and
rvlink where the paths to RV tools are relative to the script.

### Describe the reason for the change.

The current .reg files doesn't work because we can't guess the values at
build time. Without a proper installer, we can't update the files at
install time. The solution I propose is to use a batch script that
computes the path to rv and to rvpush.

### Describe what you have tested and on which operating system.

Tested on Windows 10 by dropping the new files in the etc folder of a
pre-compiled OpenRV.

### Add a list of changes, and note any that might need special
attention during the review.

Please double check the entries. I *think* I did it right, but it's
really error prone.

### If possible, provide screenshots.

---------

Signed-off-by: Kerby Geffrard <kerby.geffrard@autodesk.com>