Download Qt artifacts from the  University of California mirror (#303)
<!--
Thanks for your contribution! Please read this comment in its entirety.
It's quite important.
When a contributor merges the pull request, the title and the
description will be used to build the merge commit!

### Pull Request TITLE

It should be in the following format:

[ 12345: Summary of the changes made ] Where 12345 is the corresponding
Github Issue

OR

[ Summary of the changes made ] If it's solving something trivial, like
fixing a typo.
-->

### Linked issues
<!--
Link the Issue(s) this Pull Request is related to.

Each PR should link to at least one issue, in the form:

Use one line for each Issue. This allows auto-closing the related issue
when the fix is merged.

Fixes #12345
Fixes #54345
-->

N/A

### Summarize your change.

Avoid using Qt's main mirror since it's often really slow and
unreliable.

I got the link from https://download.qt.io/static/mirrorlist/

### Describe the reason for the change.

CI and local builds are failing because of the unreliability of the
official sources.

### Describe what you have tested and on which operating system.

Tested on MacOS 

### Add a list of changes, and note any that might need special
attention during the review.

N/A

### If possible, provide screenshots.

N/A

Signed-off-by: Kerby Geffrard <kerby.geffrard@autodesk.com>