Allow arbitrary application type and release description (#305)
<!--
Thanks for your contribution! Please read this comment in its entirety.
It's quite important.
When a contributor merges the pull request, the title and the
description will be used to build the merge commit!

### Pull Request TITLE

It should be in the following format:

[ 12345: Summary of the changes made ] Where 12345 is the corresponding
Github Issue

OR

[ Summary of the changes made ] If it's solving something trivial, like
fixing a typo.
-->

### Linked issues
<!--
Link the Issue(s) this Pull Request is related to.

Each PR should link to at least one issue, in the form:

Use one line for each Issue. This allows auto-closing the related issue
when the fix is merged.

Fixes #12345
Fixes #54345
-->

N/A

### Summarize your change.

Allows a client to set any arbitrary application release description to
OpenRV. If RELEASE, it's like before. Anything else is in red in the
About box.

Adds the application release description to RV's banner next to the
version.

### Describe the reason for the change.

It allows consumers to tag their builds to stuff like ALPHA, BETA,
"Internal Cool Studio Build".

### Describe what you have tested and on which operating system.

Tested on MacOS Ventura

### Add a list of changes, and note any that might need special
attention during the review.

N/A

### If possible, provide screenshots.

Signed-off-by: Kerby Geffrard <kerby.geffrard@autodesk.com>