Fix the build on Rocky8 (#316)
<!--
Thanks for your contribution! Please read this comment in its entirety.
It's quite important.
When a contributor merges the pull request, the title and the
description will be used to build the merge commit!

### Pull Request TITLE

It should be in the following format:

[ 12345: Summary of the changes made ] Where 12345 is the corresponding
Github Issue

OR

[ Summary of the changes made ] If it's solving something trivial, like
fixing a typo.
-->

### Linked issues
<!--
Link the Issue(s) this Pull Request is related to.

Each PR should link to at least one issue, in the form:

Use one line for each Issue. This allows auto-closing the related issue
when the fix is merged.

Fixes #12345
Fixes #54345
-->

### Summarize your change.

Uses the system OpenSSL on RHEL8 while building OpenRV.

### Describe the reason for the change.

RHEL8 uses the same version of OpenSSL than the current OpenRV, and they
enter in conflict during the build. There were a bit of documentation
about how to avoid the conflict, but this is implementing a real fix.

### Describe what you have tested and on which operating system.

Tested on Rocky8, and on Autodesk's RV CI to make sure nothing breaks
with this change.

### Add a list of changes, and note any that might need special
attention during the review.

With this change, we no longer package OpenSSL in RV if building on a
Rocky8. CentOS7 provides a openssl11 package that gives OpenSSL 1.1.1
and Rocky9 provides a compat-openssl package.

### If possible, provide screenshots.

N/A

---------

Signed-off-by: Kerby Geffrard <kerby.geffrard@autodesk.com>