Adding ALEXA ARRI LogC code (#317)
Bring back Alexa LogC and LogC Film colour transforms

### Summarize your change.

The ARRI SDK was removed when RV was open-sourced since its proprietary
code. This branch refactors the colour transforms not to use the SDK and
puts it back into the UI.

This work was started by @tbabiin but he could not complete it before
his departure, so I picked it up where he left off.

### Describe the reason for the change.

This is to restore parity with what was previously removed.

### Describe what you have tested and on which operating system.

Tested on Mac (M1) 13.6.3. 
Built on Mac 13.6.3, Windows 10, Linux Rocky 9.2 and Linux RHEL7.

Signed-off-by: Roger Nelson <roger.nelson@autodesk.com>
Co-authored-by: Inouk Babin-Ruel <inouk.babin.ruel@autodesk.com>