Disable autoEDL on RVSwitchNodes created for MMR (#320)
<!--
Thanks for your contribution! Please read this comment in its entirety.
It's quite important.
When a contributor merges the pull request, the title and the
description will be used to build the merge commit!

### Pull Request TITLE

It should be in the following format:

[ 12345: Summary of the changes made ] Where 12345 is the corresponding
Github Issue

OR

[ Summary of the changes made ] If it's solving something trivial, like
fixing a typo.
-->

### Linked issues
<!--
Link the Issue(s) this Pull Request is related to.

Each PR should link to at least one issue, in the form:

Use one line for each Issue. This allows auto-closing the related issue
when the fix is merged.

Fixes #12345
Fixes #54345
-->

N/A

### Summarize your change.

Disables the autoEDL property on the RVSwitchNode that get created as
part of Multiple Media Representation.

### Describe the reason for the change.

The default autoEDL property changes the behaviour the ImageRangeInfo()
of a RVSwitchNode

https://github.com/AcademySoftwareFoundation/OpenRV/blob/main/src/lib/ip/IPBaseNodes/SwitchIPNode.cpp#L561-L572

Where the m_offset is set based on the autoEDL property
https://github.com/AcademySoftwareFoundation/OpenRV/blob/main/src/lib/ip/IPBaseNodes/SwitchIPNode.cpp#L268

It's an issue for parent nodes that doesn't necessarily expects
RVSwitchNodes, but RVSourceNodes instead.

For example, the result of this is really different if MMR is being used
or not.

https://github.com/AcademySoftwareFoundation/OpenRV/blob/main/src/lib/ip/IPBaseNodes/SequenceIPNode.cpp#L600

FYI, the same fix is applied in the OTIO reader.
https://github.com/AcademySoftwareFoundation/OpenRV/blob/b62c9c50e4e480a818906f22fce83205fefda195/src/plugins/rv-packages/otio_reader/otio_reader.py#L487

### Describe what you have tested and on which operating system.

Tested on MacOS, but this one liner is platform agnostic.

### Add a list of changes, and note any that might need special
attention during the review.

I *think* that's the only change required, but please tell me if there's
another entrypoint for RVSwitchNodes created for MMR.

### If possible, provide screenshots.

*MMR OFF, the benchmark:*

```
reviewAppSequenceGroup_sequence : RVSequence (1)
{
    edl
    {
        int source = [ 0 1 0 ]
        int frame = [ 1 21 109 ]
        int in = [ 1001 963 0 ]
        int out = [ 1020 1050 0 ]
    }

    output
    {
        int size = [ 1920 1080 ]
        float fps = 25
        int interactiveSize = 1
        int autoSize = 1
    }

    mode
    {
        int autoEDL = 1
        int useCutInfo = 1
        int supportReversedOrderBlending = 1
    }

    composite
    {
        string inputBlendModes = [ ]
    }
}
```

*MMR ON, Without the fix:*
```
reviewAppSequenceGroup_sequence : RVSequence (1)
{
    edl
    {
        int source = [ 0 1 0 ]
        int frame = [ 1 21 109 ]
        int in = [ 1 1 0 ]
        int out = [ 20 88 0 ]
    }

    output
    {
        int size = [ 1920 1080 ]
        float fps = 25
        int interactiveSize = 1
        int autoSize = 1
    }

    mode
    {
        int autoEDL = 1
        int useCutInfo = 1
        int supportReversedOrderBlending = 1
    }

    composite
    {
        string inputBlendModes = [ ]
    }
}
```

*MMR ON, With the fix:*
```
reviewAppSequenceGroup_sequence : RVSequence (1)
{
    edl
    {
        int source = [ 0 1 0 ]
        int frame = [ 1 21 109 ]
        int in = [ 1001 963 0 ]
        int out = [ 1020 1050 0 ]
    }

    output
    {
        int size = [ 1920 1080 ]
        float fps = 25
        int interactiveSize = 1
        int autoSize = 1
    }

    mode
    {
        int autoEDL = 1
        int useCutInfo = 1
        int supportReversedOrderBlending = 1
    }

    composite
    {
        string inputBlendModes = [ ]
    }
}
```

---------

Signed-off-by: Kerby Geffrard <kerby.geffrard@autodesk.com>
Signed-off-by: Martin Chesnay <Martin.Chesnay@autodesk.com>