Base implementation of the PyMediaLibrary (#330)
Signed-off-by: Kerby Geffrard <kerby.geffrard@autodesk.com>

<!--
Thanks for your contribution! Please read this comment in its entirety.
It's quite important.
When a contributor merges the pull request, the title and the
description will be used to build the merge commit!

### Pull Request TITLE

It should be in the following format:

[ 12345: Summary of the changes made ] Where 12345 is the corresponding
Github Issue

OR

[ Summary of the changes made ] If it's solving something trivial, like
fixing a typo.
-->

### Linked issues
<!--
Link the Issue(s) this Pull Request is related to.

Each PR should link to at least one issue, in the form:

Use one line for each Issue. This allows auto-closing the related issue
when the fix is merged.

Fixes #12345
Fixes #54345
-->

N/A

### Summarize your change.

Implements the PyMediaLibrary and capabilities for MediaLibrary plugins
in OpenRV.

#### Implementation details

The plugins are lazy loaded up until a media is loaded. They are all
stored in the PyRootNode. The PyRootNode then shares the plugin object
with the children PyMediaNode. That way, they are only loaded once and
the performance stays great.

The plugins are required to implement the MediaAPI. You have an example
in the demonstration plugin.

```
{ "is_plugin_enabled" },
{ "is_library_media_url" },
{ "is_streaming" },
{ "is_redirecting" },
{ "get_http_cookies" },
{ "get_http_headers" },
{ "get_http_redirection" }
```

#### Demonstration plugin
Please, don't use it in production. 

It comes with a demonstration plugin that needs to be enabled with the
`RV_MEDIA_LIBRARY_UNSECURE_DEMO_ENABLED` set to 1.

Once enabled, you can set the cookies for one site using
`RV_MEDIA_LIBRARY_UNSECURE_DEMO_COOKIES`. It's parsed by Python's
SimpleCookie module and it allows you to set multiple cookies for one
site like this `cookie1=value1; [Domain=domain1; Path=path1;]
cookie2=value2; [Domain=domain1; Path=path2;]`

You can also set headers with `RV_MEDIA_LIBRARY_UNSECURE_DEMO_HEADERS`.
The format is the standard header format `key1=value1;key2=value2`

Finally, you can tell the demo plugin to download the file on disk and
redirect the FileSourceIPNode to it with
`RV_MEDIA_LIBRARY_UNSECURE_DEMO_FORCE_LOCAL_CACHE` set to 1.
 
#### Debugging the plugins

You can debug what gets parsed and sent to ffmpeg with
`RV_DEBUG_FFMPEG_COOKIES` and `RV_DEBUG_FFMPEG_HEADERS`.


#### EventMediaLibrary

The EventMediaLibrary got migrated to a plugin so we don't have two
implementations to maintain.

### Describe the reason for the change.

RVIO cannot use the EventMediaLibrary because RV Modes are not loaded.
If your integration relies on being able to provide cookies or headers
to RV/RVIO to play and export from media stored on the web, it only
works in RV, and not in RVIO.

The MediaLibrary plugins are always loaded, unless you set
`RV_USE_MEDIA_LIBRARY_PLUGINS` to 0.

### Describe what you have tested and on which operating system.

I tested it on MacOS, but nothing is really platform-specific. 

### Add a list of changes, and note any needing special attention during
the review.

See pull request comments. 

### If possible, provide screenshots.

N/A

---------

Signed-off-by: Kerby Geffrard <kerby.geffrard@autodesk.com>